[ADD] Alquiler (REST)

+++ coca-tiendaProyecto/src/main/java/com/coca/tienda/repositories/MaterialRepository.java
@@ -3,6 +3,6 @@ package com.coca.tienda.repositories;
 import com.coca.tienda.entities.Material;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialRepository extends JpaRepository<Material, Long> {
+public interface MaterialRepository extends JpaRepository<Material, Integer> {
 	
 }
